lego_scrapper/requirements.txt
annotated-types==0.7.0
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.3
idna==3.10
pydantic==2.11.10
pydantic_core==2.33.2
PyYAML==6.0.3
requests==2.32.5
soupsieve==2.8
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
